AlembicStreamDescriptor is internal again and all code is moved to the StreamPlayer.

+++ com.unity.formats.alembic/Editor/Importer/AlembicImporter.cs
@@ -254,8 +254,8 @@ namespace UnityEditor.Formats.Alembic.Importer
                     frames[0].value = 0.0f;
                     frames[0].time = 0.0f;
                     frames[0].outTangent = 1.0f;
-                    frames[1].value = (float)streamDescr.mediaDuration;
-                    frames[1].time = (float)streamDescr.mediaDuration;
+                    frames[1].value = streamDescr.mediaDuration;
+                    frames[1].time = streamDescr.mediaDuration;
                     frames[1].inTangent = 1.0f;
 
                     var curve = new AnimationCurve(frames);
